Inline QILLR gates into QPU circuits
