+++ lib/Dialect/QILLR/CMakeLists.txt
@@ -1,3 +1,2 @@
-add_subdirectory(Extensions)
 add_subdirectory(IR)
 add_subdirectory(Transforms)
