+++ lib/Dialect/QILLR/IR/QILLRBase.cpp
@@ -20,8 +20,6 @@ using namespace mlir::quantum;
 
 //===----------------------------------------------------------------------===//
 
-namespace {} // namespace
-
 //===----------------------------------------------------------------------===//
 // QILLRDialect
 //===----------------------------------------------------------------------===//
