+++ lib/Dialect/QPU/IR/QPUOps.cpp
@@ -6,6 +6,7 @@
 #include "quantum-mlir/Dialect/QPU/IR/QPUOps.h"
 
 #include "mlir/Interfaces/FunctionImplementation.h"
+#include "mlir/Transforms/InliningUtils.h"
 #include "quantum-mlir/Dialect/QPU/IR/QPUAttributes.h"
 #include "quantum-mlir/Dialect/QPU/IR/QPUTypes.h"
 
@@ -35,6 +36,54 @@
 
 using namespace mlir;
 using namespace mlir::qpu;
+
+//===----------------------------------------------------------------------===//
+// QPUDialect Interfaces
+//===----------------------------------------------------------------------===//
+
+namespace {
+/// This class defines the interface for handling inlining with gate operations.
+struct QPUInlinerInterface : public DialectInlinerInterface {
+    using DialectInlinerInterface::DialectInlinerInterface;
+
+    //===--------------------------------------------------------------------===//
+    // Analysis Hooks
+    //===--------------------------------------------------------------------===//
+
+    /// Call operations can always be inlined
+    bool isLegalToInline(
+        Operation* call,
+        Operation* callable,
+        bool wouldBeCloned) const final
+    {
+        return true;
+    }
+
+    /// All operations can be inlined.
+    bool isLegalToInline(Operation*, Region*, bool, IRMapping &) const final
+    {
+        return true;
+    }
+
+    /// All gate bodies can be inlined.
+    bool isLegalToInline(Region*, Region*, bool, IRMapping &) const final
+    {
+        return true;
+    }
+
+    //===--------------------------------------------------------------------===//
+    // Transformation Hooks
+    //===--------------------------------------------------------------------===//
+
+    /// Handle the given inlined terminator by replacing it with a new operation
+    /// as necessary.
+    void handleTerminator(Operation* op, Block* newDest) const final {};
+
+    /// Handle the given inlined terminator by replacing its operands
+    void handleTerminator(Operation* op, ValueRange valuesToRepl) const final {
+    };
+};
+} // namespace
 
 //===----------------------------------------------------------------------===//
 // ExecuteOp
@@ -173,4 +222,5 @@ void QPUDialect::registerOps()
 #define GET_OP_LIST
 #include "quantum-mlir/Dialect/QPU/IR/QPUOps.cpp.inc"
         >();
+    addInterfaces<QPUInlinerInterface>();
 }
